MCS51的代码仓库重建完毕

+++ 3_LED_Segment_Displays/USER/1seg_74xx164_main.c
@@ -10,7 +10,7 @@
 #define u16  unsigned int
 #endif
 
-void Delay(uint t)
+void Delay(u16 t)
 {
 	while(t--);
 }
